+++ services/services/screen_capture/ipc/screen_capture_service_stub.cpp
@@ -273,9 +273,9 @@ int32_t ScreenCaptureServiceStub::ExcludeContent(MessageParcel &data, MessagePar
     CHECK_AND_RETURN_RET_LOG(windowIdSize < MAX_FILTER_CONTENTS_COUNT, MSERR_INVALID_STATE,
                              "windowID size is exceed max range");
     if (windowIdSize > 0) {
-        std::vector<int> vec;
+        std::vector<uint64_t> vec;
         for (int32_t i = 0; i < windowIdSize; i++) {
-            vec.push_back(data.ReadInt32());
+            vec.push_back(data.ReadUint64());
         }
         contentFilter.windowIDsVec = vec;
     }
